refactoring preprocessing of dataset
