+++ modules/gans/__init__.py
@@ -1,0 +1,3 @@
+from .dcgan import DCGAN
+from .srgan import SRGAN
+from .wasserstein_gan import WGANGC
