+++ modules/gans/wasserstein_gan.py
@@ -20,10 +20,10 @@ from statistics import mean
 import imagesize
 from multiprocessing.pool import ThreadPool
 
-from modules.batch_maker import BatchMaker
-from modules.models import discriminator_models_spreadsheet, generator_models_spreadsheet
-from modules.keras_extensions.custom_tensorboard import TensorBoardCustom
-from modules.helpers import time_to_format, get_paths_of_files_from_path
+from ..batch_maker import BatchMaker
+from ..models import discriminator_models_spreadsheet, generator_models_spreadsheet
+from ..keras_extensions.custom_tensorboard import TensorBoardCustom
+from ..helpers import time_to_format, get_paths_of_files_from_path
 
 # Custom loss function
 def wasserstein_loss(y_true, y_pred):
