+++ modules/models/generator_models_spreadsheet.py
@@ -3,8 +3,8 @@ from keras.initializers import Initializer, RandomNormal
 from keras.layers import Layer, Dense, Reshape, Conv2D, LeakyReLU, Add, Lambda
 from tensorflow import Tensor
 
-from modules.models.custom_layers import deconv_layer, conv_layer, RRDB1
-from modules.helpers import count_upscaling_start_size
+from .custom_layers import deconv_layer, conv_layer, RRDB1
+from ..helpers import count_upscaling_start_size
 
 def mod_testing(inp:Union[Tensor, Layer], image_shape:tuple, image_channels:int, kernel_initializer:Initializer=RandomNormal(stddev=0.02)):
   st_s = count_upscaling_start_size(image_shape, 3)
