+++ train_srgan.py
@@ -26,7 +26,7 @@ if gpus:
 
 from keras.optimizers import Adam
 
-from modules.gans.srgan import SRGAN
+from modules.gans import SRGAN
 from modules.keras_extensions.custom_adam_optimizer import AccumulateAdam
 from settings.srgan_settings import *
